OWB-184 add BeanManager as managed bean as soon as possible
The problem which caused almost all TCKs to got skipped was that the BeanManager
got only added with configureDefaultBeans() at the end of the container startup.
If there was a DeploymentException or any similar unexpedted behaviour detected,
the BeanManager never got added as managed bean, thus causing an Exception
while firing the ShutdownEvent.


git-svn-id: https://svn.apache.org/repos/asf/incubator/openwebbeans/trunk@886349 13f79535-47bb-0310-9956-ffa450edef68