OWB-414 drop doPrivileged from ClassLoader detection
retrieving the CurrentContextClassLoader via SecurityManager
is a highly expensive task. Doing this in the most heavy spot
of OWB means a measured performance penalty from 20 to 30%.
Given the fact that OWB (even if running in Geronimo or any 
other multi ClassLoader scenario) will always get bootstrapped
in the context of a WebApp, we should not get any problems
by dropping it. 


git-svn-id: https://svn.apache.org/repos/asf/openwebbeans/trunk@966924 13f79535-47bb-0310-9956-ffa450edef68