OWB-419 fix our ELResolvers
This commit cleans up a lot old stuff and improves the caching
for EL evaluations heavily. We now cache every resolved Bean proxy
until the end of the request. This doesn't hurt even if the managed beans
have a shorter lifecycle, because the proxy will take care of such scenarios!


git-svn-id: https://svn.apache.org/repos/asf/openwebbeans/trunk@966925 13f79535-47bb-0310-9956-ffa450edef68