fix EmptyAnnotationLiteral annotationType detection when used with RI-API
The CDI RI API uses the annotationType() method already in it's constructor. 
As we had to implement the logic again in our own class (due to a bug 
in an older RI API version), this now happens before our own ct has a chance
to do it's work. Thus we need to lazily resolve the annotationType.


git-svn-id: https://svn.apache.org/repos/asf/openwebbeans/trunk@1806390 13f79535-47bb-0310-9956-ffa450edef68