OWB-1428 make default bean-discovery-mode configurable
There was a really problematic change in the CDI-4.0 specification which will break many applications. They switched the bean-discovery-mode of an empty beans.xml file (or a beans.xml without any version) from ALL to ANNOTATED (Despite warnings that his is totally backward incompatible and could easily have been avoided).

The default in OWB is still ALL, but it can be configured to any other bean-discovery-mode with this config switch to pass this very TCK test.